requirement.txt
da-vinci==0.2.2
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.1.7
django-ajaximage==0.2.8
django-appconf==1.0.3
django-crispy-forms==1.7.2
django-formtools==2.1
django-imagekit==4.0.2
django-import-export==1.2.0
django-js-asset==1.2.2
django-mptt==0.9.1
django-orderable==6.0.1
django-respite==1.4
django-rest-framework==0.1.0
django-reversion==3.0.3
django-simpleui==1.4.7
django-stdimage==4.0.1
django-thumbnails==0.2.1
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
djangorestframework-stubs==0.3.0
easy-thumbnails==2.6
et-xmlfile==1.0.1
future==0.17.1
httplib2==0.12.1
jdcal==1.4
jieba==0.39
odfpy==1.4.0
openpyxl==2.6.0
pilkit==2.0
Pillow==5.4.1
progressbar2==3.39.2
psycopg2==2.7.7
PyJWT==1.7.1
python-utils==2.3.0
pytz==2018.9
PyYAML==3.13
shortuuid==0.5.0
six==1.12.0
tablib==0.12.1
unicodecsv==0.14.1
xlrd==1.2.0
XlsxWriter==1.1.5
xlwt==1.3.0
